chore(tests): split up application submission e2e tests
