+++ .github/workflows/test.yml
@@ -192,17 +192,18 @@ jobs:
       matrix:
         spec:
           [
-            "/*.spec.js",
-            "account/create/**/*.spec.js",
-            "account/manage/**/*.spec.js",
-            "account/password-reset/**/*.spec.js",
-            "account/reactivated/**/*.spec.js",
-            "account/sign-in/**/*.spec.js",
-            "account/sign-out/**/*.spec.js",
-            "account/suspended/**/*.spec.js",
-            "application-submission/*.spec.js",
-            "application-submission/multiple-policy-type/**/*.spec.js",
-            "application-submission/single-policy-type/**/*.spec.js",
+            "application-submission/multiple-policy-type/*.spec.js",
+            "application-submission/multiple-policy-type/business-conditions/*.spec.js",
+            "application-submission/multiple-policy-type/buyer-conditions/*.spec.js",
+            "application-submission/multiple-policy-type/buyer-conditions/traded-with-buyer/*.spec.js",
+            "application-submission/multiple-policy-type/export-contract-conditions/*.spec.js",
+            "application-submission/multiple-policy-type/policy-contract-conditions/*.spec.js",
+            "application-submission/single-policy-type/*.spec.js",
+            "application-submission/single-policy-type/business-conditions/*.spec.js",
+            "application-submission/single-policy-type/buyer-conditions/*.spec.js",
+            "application-submission/single-policy-type/buyer-conditions/traded-with-buyer/*.spec.js",
+            "application-submission/single-policy-type/export-contract-conditions/*.spec.js",
+            "application-submission/single-policy-type/policy-contract-conditions/*.spec.js",
             "cannot-skip-flow/**/*.spec.js",
             "check-your-answers/export-contract/*.spec.js",
             "check-your-answers/export-contract/change-your-answers/about-goods-or-services/*.spec.js",
