+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/multiple-policy-type/business-conditions/submit-an-application-multiple-policy-type-with-different-trading-address.spec.js
@@ -1,0 +1,29 @@
+import { APPLICATION } from '../../../../../../../constants';
+
+context(
+  'Insurance - submit an application - Multiple policy type, different trading address - As an Exporter, I want to submit my completed credit insurance application, So that UKEF can process and make a decision on my application',
+  () => {
+    let referenceNumber;
+
+    before(() => {
+      cy.completeSignInAndSubmitAnApplication({
+        policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
+        differentTradingAddress: true,
+      }).then((refNumber) => {
+        referenceNumber = refNumber;
+      });
+    });
+
+    beforeEach(() => {
+      cy.saveSession();
+    });
+
+    after(() => {
+      cy.deleteApplication(referenceNumber);
+    });
+
+    it('should successfully submit the application and redirect to `application submitted`', () => {
+      cy.assertApplicationSubmittedUrl(referenceNumber);
+    });
+  },
+);
