+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/multiple-policy-type/business-conditions/submit-an-application-multiple-policy-type-with-different-trading-name-address-and-broker.spec.js
@@ -1,7 +1,7 @@
-import { APPLICATION } from '../../../../../../constants';
+import { APPLICATION } from '../../../../../../../constants';
 
 context(
-  'Insurance - submit an application - Multiple policy type, different trading name and address - As an Exporter, I want to submit my completed credit insurance application, So that UKEF can process and make a decision on my application',
+  'Insurance - submit an application - Single policy type with a different trading name, address and broker - As an Exporter, I want to submit my completed credit insurance application, So that UKEF can process and make a decision on my application',
   () => {
     let referenceNumber;
 
@@ -10,6 +10,7 @@ context(
         policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
         differentTradingName: true,
         differentTradingAddress: true,
+        usingBroker: true,
       }).then((refNumber) => {
         referenceNumber = refNumber;
       });
