+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/multiple-policy-type/export-contract-conditions/submit-an-application-multiple-policy-type-attempted-private-market-cover-and-using-an-agent-no-charges.spec.js
@@ -1,13 +1,15 @@
-import { APPLICATION } from '../../../../../../constants';
+import { APPLICATION } from '../../../../../../../constants';
 
-context('Insurance - submit an application - Single policy type, Loss payee - Financial details UK ', () => {
+context('Insurance - submit an application - Multiple policy type, attempted private market cover, using an agent, no agent charges', () => {
   let referenceNumber;
 
   before(() => {
     cy.completeSignInAndSubmitAnApplication({
       policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
-      isAppointingLossPayee: true,
-      lossPayeeIsLocatedInUK: true,
+      totalContractValueOverThreshold: true,
+      attemptedPrivateMarketCover: true,
+      isUsingAgent: true,
+      agentIsCharging: false,
     }).then((refNumber) => {
       referenceNumber = refNumber;
     });
