+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/multiple-policy-type/export-contract-conditions/submit-an-application-multiple-policy-type-attempted-private-market-cover-and-using-an-agent-with-charges-fixed-sum.spec.js
@@ -1,16 +1,18 @@
-import { APPLICATION } from '../../../../../../constants';
+import { APPLICATION } from '../../../../../../../constants';
 
 context(
-  'Insurance - submit an application - Single policy type with a different trading name, address and broker - As an Exporter, I want to submit my completed credit insurance application, So that UKEF can process and make a decision on my application',
+  'Insurance - submit an application - Multiple policy type, attempted private market cover, using an agent, agent is charging - fixed sum method',
   () => {
     let referenceNumber;
 
     before(() => {
       cy.completeSignInAndSubmitAnApplication({
         policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
-        differentTradingName: true,
-        differentTradingAddress: true,
-        usingBroker: true,
+        totalContractValueOverThreshold: true,
+        attemptedPrivateMarketCover: true,
+        isUsingAgent: true,
+        agentIsCharging: true,
+        agentChargeMethodFixedSum: true,
       }).then((refNumber) => {
         referenceNumber = refNumber;
       });
