+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/multiple-policy-type/export-contract-conditions/submit-an-application-multiple-policy-type-attempted-private-market-cover.spec.js
@@ -1,9 +1,13 @@
-context('Insurance - submit an application - Single policy type, no broker - As an Exporter, I want to submit my completed credit insurance application with a value greater than 500000, So that UKEF can process and make a decision on my application', () => {
+import { APPLICATION } from '../../../../../../../constants';
+
+context('Insurance - submit an application - Multiple policy type, attempted private market cover', () => {
   let referenceNumber;
 
   before(() => {
     cy.completeSignInAndSubmitAnApplication({
-      policyValueOverMvpMaximum: true,
+      policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
+      totalContractValueOverThreshold: true,
+      attemptedPrivateMarketCover: true,
     }).then((refNumber) => {
       referenceNumber = refNumber;
     });
