+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/multiple-policy-type/policy-conditions/submit-an-application-multiple-policy-type-loss-payee-uk.spec.js
@@ -1,16 +1,13 @@
-import { APPLICATION } from '../../../../../../constants';
+import { APPLICATION } from '../../../../../../../constants';
 
-context('Insurance - submit an application - Multiple policy type, attempted private market cover, using an agent, agent is charging - percentage method', () => {
+context('Insurance - submit an application - Single policy type, Loss payee - Financial details UK ', () => {
   let referenceNumber;
 
   before(() => {
     cy.completeSignInAndSubmitAnApplication({
       policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
-      totalContractValueOverThreshold: true,
-      attemptedPrivateMarketCover: true,
-      isUsingAgent: true,
-      agentIsCharging: true,
-      agentChargeMethodPercentage: true,
+      isAppointingLossPayee: true,
+      lossPayeeIsLocatedInUK: true,
     }).then((refNumber) => {
       referenceNumber = refNumber;
     });
