--- e2e-tests/insurance/cypress/e2e/journeys/application-submission/multiple-policy-type/submit-an-application-multiple-policy-type-attempted-private-market-cover-and-using-an-agent-with-charges-fixed-sum.spec.js
@@ -1,30 +1,0 @@
-import { APPLICATION } from '../../../../../../constants';
-
-context('Insurance - submit an application - Multiple policy type, attempted private market cover, using an agent, agent is charging - fixed sum method', () => {
-  let referenceNumber;
-
-  before(() => {
-    cy.completeSignInAndSubmitAnApplication({
-      policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
-      totalContractValueOverThreshold: true,
-      attemptedPrivateMarketCover: true,
-      isUsingAgent: true,
-      agentIsCharging: true,
-      agentChargeMethodFixedSum: true,
-    }).then((refNumber) => {
-      referenceNumber = refNumber;
-    });
-  });
-
-  beforeEach(() => {
-    cy.saveSession();
-  });
-
-  after(() => {
-    cy.deleteApplication(referenceNumber);
-  });
-
-  it('should successfully submit the application and redirect to `application submitted`', () => {
-    cy.assertApplicationSubmittedUrl(referenceNumber);
-  });
-});
