--- e2e-tests/insurance/cypress/e2e/journeys/application-submission/multiple-policy-type/submit-an-application-multiple-policy-type-attempted-private-market-cover.spec.js
@@ -1,27 +1,0 @@
-import { APPLICATION } from '../../../../../../constants';
-
-context('Insurance - submit an application - Multiple policy type, attempted private market cover', () => {
-  let referenceNumber;
-
-  before(() => {
-    cy.completeSignInAndSubmitAnApplication({
-      policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
-      totalContractValueOverThreshold: true,
-      attemptedPrivateMarketCover: true,
-    }).then((refNumber) => {
-      referenceNumber = refNumber;
-    });
-  });
-
-  beforeEach(() => {
-    cy.saveSession();
-  });
-
-  after(() => {
-    cy.deleteApplication(referenceNumber);
-  });
-
-  it('should successfully submit the application and redirect to `application submitted`', () => {
-    cy.assertApplicationSubmittedUrl(referenceNumber);
-  });
-});
