--- e2e-tests/insurance/cypress/e2e/journeys/application-submission/multiple-policy-type/submit-an-application-multiple-policy-type-using-an-agent-cover.spec.js
@@ -1,26 +1,0 @@
-import { APPLICATION } from '../../../../../../constants';
-
-context('Insurance - submit an application - Multiple policy type, using an agent', () => {
-  let referenceNumber;
-
-  before(() => {
-    cy.completeSignInAndSubmitAnApplication({
-      policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
-      isUsingAgent: true,
-    }).then((refNumber) => {
-      referenceNumber = refNumber;
-    });
-  });
-
-  beforeEach(() => {
-    cy.saveSession();
-  });
-
-  after(() => {
-    cy.deleteApplication(referenceNumber);
-  });
-
-  it('should successfully submit the application and redirect to `application submitted`', () => {
-    cy.assertApplicationSubmittedUrl(referenceNumber);
-  });
-});
