+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/single-policy-type/business-conditions/submit-an-application-single-policy-type-with-different-trading-address.spec.js
