+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/single-policy-type/export-contract-conditions/submit-an-application-single-policy-type-attempted-private-market-cover-and-using-an-agent-with-charges-percentage.spec.js
