+++ e2e-tests/insurance/cypress/e2e/journeys/application-submission/single-policy-type/policy-conditions/submit-an-application-single-policy-type-maximum-cover.spec.js
@@ -1,15 +1,11 @@
-import { APPLICATION } from '../../../../../../constants';
-
 context(
-  'Insurance - submit an application - Single policy type with a different trading name and broker - As an Exporter, I want to submit my completed credit insurance application, So that UKEF can process and make a decision on my application',
+  'Insurance - submit an application - Single policy type, no broker - As an Exporter, I want to submit my completed credit insurance application with a value greater than 500000, So that UKEF can process and make a decision on my application',
   () => {
     let referenceNumber;
 
     before(() => {
       cy.completeSignInAndSubmitAnApplication({
-        policyType: APPLICATION.POLICY_TYPE.MULTIPLE,
-        differentTradingName: true,
-        usingBroker: true,
+        policyValueOverMvpMaximum: true,
       }).then((refNumber) => {
         referenceNumber = refNumber;
       });
